replace bitcast
